Use "snazzy" colorscheme in MATE Terminal
Or an edited version of it; it's nice and lively and visible but I
prefer the bg to be less blue.

The original scheme is https://github.com/sindresorhus/hyper-snazzy,
but I used https://github.com/miedzinski/konsole-snazzy as the basis
for color values.

I'm going to shop around a bit more for color schemes but this one is
good at the moment.