Install fzf
Before everyone calls in to tell me ripgrep is better than ag... I
know.  But it's not in Ubuntu repos, and I've had problems installing
ripgrep from the downloadable binaries (missing an instruction set in
the CPU on my work PC).  So it's unfortunately not the easiest thing
to get working, and ag seems to perform well enough for my purposes.