Remove antigen zsh setup in favor of prezto
oh-my-zsh is easy as heck, but its performance was kind of wack; five
seconds for a shell to pop up is definitely both a first world
problem, but pretty slow regardless.  I'd still suggest newcomers to
zsh use OMZ to start off, so long as they know the downsides.

antigen did not seem to fix this problem.  Maybe I was using it
wrong.  Furthermore, it was really flaky with the themes I was trying
to use, and theme switching was just flat out broken.

prezto is much faster, has some nice themes, is pretty easy to fork
and hack on yourself...

The only problems I've ran into is regarding some of the syntax
highlighting, which is optional, and also fzf didn't work as smoothly
as I hoped.  I'll still try to get it in there, but I need to focus on
other things ASAP.