Install MPD... mostly
I'm not even using the MATE part of this setup anymore[1] but you will
probably need to add mpd to MATE's startup applications to get it to
do anything.

Removed clementine from the setup, because... we now have mpd.

The one downside is no notifications.  I found some rando code to do
this but am planning on writing my own attempt at some point as well.

[1] Switched to StumpWM, which I'm working on adding to this setup as
    well.  But I probably won't get rid of the MATE stuff either,
    because it makes a fine fallback if I ever do run into problems
    with Stump, and it pulls in various apps that can be run from
    Stump easily.