Set up lightdm-gtk-greeter
And this is the 'better' greeter.

There is also lightdm-webkit-greeter, which is prettier yet also more
functional than the 'slick greeter' which is MATE's default.  What I
don't like about it is that it... pulls in all of Webkit.  Which makes
me leery from a security/attack-surface perspective, given that this
is literally *the password screen to your computer*.

So I'll probably rough it out with this one.  You only see it for a
minute tops as you log into the machine, really.