Unbound MATE media keys, replaced with MPD
I'm liking MPD more and more, to the point where I don't really want
to use other media players.  Furthermore, I only really use my media
keys to control MPD.  Finally, I'm finding it difficult to expose the
right D-Bus interfaces to allow MPD to pick up MATE's media key
signals that it emits by default.  MPRIS and playerctl would be an
option if an MPD MPRIS daemon and playerctl were in Ubuntu repos.
They are not (at least for 18.04), and playerctl is not so easy to
compile under 18.04, so...

We go for the next best thing, what I have done in the past to get
media keys and MPD working.