VSCode settings
No, not getting rid of emacs.

The problem with emacs is that sometimes you want the shiny stuff.  You
don't want to have to think about extensions too hard.  You don't want
to write any more dang lisp.  VSCode comes in handy in those cases.

In particular, I *still* deal with a lot of mixed-mode PHP that nothing
in emacs handles very well.  The only thing that does it well, is Vim
and VSCode.  I'd rather use the latter.

VSCode, for as nice as it is, will never have anything like org-mode,
nor will it have anything like magit.  Those argue against yeeting
emacs.

This will set up a pretty minimal VSCode, with telemetry turned off and
with a few custom file associations, Pragmata as the font, and one
theme I liked.  VSCode is pretty good at offering relevant plugins for
whatever language you're in, so if I like something I'll probably add
that to the mix as well.