Change org-capture bind, add open-junk-file bind
My window manager is slowly becoming a mechanism to open emacs frames
with :P

open-junk-file is a plugin that lets you create "junk" files.  I don't
know about you, but I know I create a ton of "test" files all over my
hard drive; this intends to slow that down a bit.

I put this on S-s-j, and moved org-capture (which was there) to S-s-o.