Remove foreign language support
By default, check-language-support pulls in fcitx, which is hopelessly
broken on Ubuntu MATE 18.04 (ex. you can't click the applet to change
stuff, because it'll friggin freeze the desktop until you kill
fcitx[1]).  I could figure out what ibus packages are needed and
all... or I could just type the English I want into Google Translate
and copypasta the Japanese out of it, for as often as I need it.

[1] https://bugs.launchpad.net/ubuntu/+source/fcitx/+bug/1767765