Switch between workspaces on same monitor
i3 creates one initial workspace per each monitor, with the primary
taking workspace 1.  (So your second monitor would get 2, and so on)
Then additional workspaces are created depending on what monitor you
were on when you created it.

So if you work in a multi-monitor environment, "workspace next" jumps
between your monitors, which is disconcerting.  This fixes that.

I'm still working on a strategy for i3.  The way i3 handles different
monitors is by no means wrong, and is probably easy to conceptualize,
implement, and maintain.  (Like a lot of things in i3, actually, and
this is very refreshing.)  But I'm very strongly used to having some
work on workspace 1, then some work on workspace 2, so on... and this
breaks this somewhat.  I like how i3 does it because I can now
semi-permanently park a terminal on one monitor, which required making
it sticky under MATE, so I don't wish to break it too much :P