Change i3lock again, use tiling image
This will make it easier to handle multimonitor, I don't really want
to futz with i3lock-fancy and friends.

I don't know the license of the tiling image.  It may well be unfree.
I found it in someone's dump of tiling wallpapers[1].  I reverse image
searched and didn't see anyone claiming copyright, and then took the
original and modified the contrast and saturation so that the monitor
doesn't make my eyes scream in a dark room.

[1] https://imgur.com/gallery/BL5Rs