.tmux.conf assorted changes
tmux has a lot more "stuff" in it than screen does and that means it's
kind of hairy to configure but does a lot more stuff.

 - C-o Space now switches to the previously used window.  I wish I
   could have bound C-o S-Tab but... for hysterical reasons, S-Tab is
   something of a taboo configuration and I'm not interested in
   throwing out a generation of VTxxx emulators to get there

 - Some keys now let you repeat them, e.g. C-o JJJ... will keep moving
   the window down, and you can even KJJJK if you want to adjust

 - Remove the black status bar, remove background from prev/current
   window statuses, and made the colors brighter.  It doesn't
   perfectly fit in with the rest of my desktop, but the whole thing
   is kind of mismatched vaporwave anyways sooooooo

Ultimately, I'd like the bindings to be congruent with my i3 bindings.
The problem is that for various reasons (use a small monitor for
terminal use at work) I'm much more likely to futz with split widths
inside tmux and with i3!window/tmux!pane positioning inside i3... as a
result i3 gets easier window positioning tools, and tmux gets easier
split-moving tools.