Changed i3lock wallpaper
This one has the advantage that I made it, and so I don't have to
worry about licensing.

It was simple as heck too; in GIMP, make a 50x50 image, create some
noise, scale it 20x WITHOUT filtering, and then add a 1px white grid
at the edges of the "pixels", and then colorize to taste.  You're
done!