Moved icons into the scripts
This has the advantage that the icons can actually change
corresponding to script values, but only some of the scripts actually
use this.  I'll come back around and fix things later.

I don't really know Perl and am on a machine without a battery so I
can't actually test the battery script and am fairly sure there's
problems with it that I'll need to iron out on a laptop...