Display weather status in i3blocks using wttr.in
This should set it up where wttr.in is being fetched hourly via
cron (wouldn't recommend doing it more frequently) and cached locally.

Right now I'm only pulling current conditions and temperature.

By default wttr.in geolocates but the geolocation can be fiddly as per
usual.  At work, it thinks I'm at home.  At home it thinks I'm in a
city named the same as the city I work in, but is about eight hours to
the north of me and can have very different weather from here.  As
such, you may have to manipulate the script from time to time to do
what is needed...