Replace wttr.in with crazy homebrew crap
wttr.in has been hitting a lot of rough patches lately, with it
running out of API limits and scalability issues.

I looked around and cooked up the crazy junk I have at
https://github.com/jws85/scripts.  I'm not going to claim the code is
good or anything.  But it is the bare minimum to get something to show
up on my WM bar.

Note that:

 - due to the downsides, the needed cron jobs (wx-fetch and
   potentially geo-fetch) are not set up yet
 - you'll need an OpenWeatherMap API key
 - you'll need to figure out your location, or consent to having
   your geolocation data being sent over HTTP (!!)