Site-customizable .config/i3/config
This is something I've been wanting for a second or two -- mostly to
fiddle around with colors and workspace icons, but still.

genconfig is a Python script that interpolates config.json's data
into config.j2 via Jinja2[1].  It does the minimum possible thing to
get this working, and is probably otherwise pretty terrible code.

We pull in python3-jinja2 via apt for this express purpose.

[1] https://jinja.palletsprojects.com/en/2.11.x/