Remove unattended-upgrades
Yes yes security blah blah DARKSIDED blah blah blah BLAH

Do any of you securityniks understand how ANNOYING it is to be doing
stuff in Firefox and then you're met with the stupid sad dinosaur
error[1] and now you have to EXIT OUT OF FIREFOX AND WAIT FOR IT TO
RESTART?

We're all at home using our devices to do things through web browsers.
What if this happened in the middle of a telehealth session?  What if
it happened in the middle of an important client presentation?

Yeah I could learn how to configure this package or something but you
know what, I don't care that much, since I manually run updates fairly
often (every week or two, or if I hear there's an important fix to
something on the level of heartbleed/shellshock).  And you know what?
IT'S MY GOSH DARN CHOICE TO DO SO.

At this point, I use Windows as a Wintendo, so I don't care so much
about its firehose of updates.  (Except when I want to play
something.)  But it's frustrating beyond belief on a workstation OS.

The only place I could possibly understand unattended-upgrades-like
utilities would be on something like a firewall or a publicly visible
box, where it'd be nice to get automatic security updates.

[1] https://user-images.githubusercontent.com/3396686/76416764-fb92bd00-639b-11ea-853c-2290d9bb496f.png