Changed active/focus colors
The active color is now the former focus color -- a blue.  I liked the
blue, but it blends in with the color of my active Firefox tab, making
it hard to know which is active at a glance.  The focus color is now a
purple.  The old active color was IN YOUR FACE orange that was
somewhat distracting in a multimonitor setup.

The purple is nice but almost too subtle.  I may have to change again.