Allow for *some* lockscreen configurability
I didn't like setting the lockscreen on boot because i3lock has
annoying rules -- the image MUST be the exact same size as the screen,
and it MUST be a PNG.

I tried betterlockscreen[1] which is pretty and works "better" but I
don't want to be cloning everyone's git repo and putting it in place
and was hoping to put it in a .deb.  Well that's not easy either.
To put stuff in a PPA, you have to dot all these i's and cross these
t's and man I just want to put a package somewhere that I can apt
install without making sure the package works on EVERY SINGLE DEBIAN
COMPUTER THAT MIGHT EVER EXIST I'm not interested in being a packager,
I just want to install stuff maaaaaaannnnn

So this is the easiest way to get from point A to point B, at the cost
of some of the niceties of betterlockscreen.  By default we show a
flat color (it's actually nicely reminiscent of Windows 95 somehow)
and if one should want to, they can do a

  convert img.jpg -resize [dims] -format png ~/.i3lock

and then add the utilities.i3lock option to "-i ~/.i3lock" and reload.

[1] https://github.com/pavanjadhaw/betterlockscreen