Split server role into docker/kvm roles
There is a very real chance "dev-desktop" becomes "dev-laptop" and
"full-desktop" becomes "battlestation" or something and learns KVM...
I'd love to make a real boss machine and run KVM VFIO on it.

Docker is much easier to work with on laptops, which is why it is
still installed 'everywhere'

Also removed whatever software was there that allowed "virtual
networking" -- it never got used.  It just didn't spark joy, sorry!