Including Ansible dependencies
This is needed to install Ansible 2.10.x, which I guess changed their
packaging...

...and I was trying to update Ansible to fix some dumb systemd bug
which is some weird interaction with the kernel, and the fix is to get
systemd 245.6 (I'm on .4) or downgrade Linux to a 5.7 kernel
(on 5.8.0-7625).  Probably going to have to pin the kernel until the
bug is fixed.