...and for Alacritty
I need to wire up alacritty (or better yet, the underlying X font
stuff) so that the default font is Fira Code, and then I can come
around and stick the fancy font I spent money on in manually.

Apparently this can be done by setting the font to monospace
everywhere, and then I'd set monospace to Pragmata Pro myself and it
should "take" in all the applications.  Past experience has been iffy;
the font face may be right, but often the font sizes are all wrong.