Change default font to 11pt Fira Code
This is to have a 'sane' font available to bootstrap from.

I set Fira Code as 'monospace' using fonts.conf, so it can be "easily"
swapped out later on, and then set the font configuration of i3,
alacritty, etc to 'monospace'.  This gets me consistency, but it gets
me the ability to easily swap out this font for something else.  I own
a copy of the regular weight of PragmataPro (last I checked, about
$70/weight) and would like it everywhere, but changing all the
settings everywhere is a pain in the keister.

If you should wish to change the font, edit
~/.config/fontconfig/fonts.conf and change the <string> value under
<edit binding> to the name of your font.  The font's name can be found
with an 'fc-list'.

You will need to restart whatever programs for the setting to 'take'.

---

I hate font configuration, by the way.  No one can agree on proper
sizing, and you have to know some font designer bullsh!t to get very
far.

By default, alacritty tries to use your monitor's DPI to size it to an
actual physical size rather than points/pixels.  This is valiant but
foolhardy; thankfully there is a setting to remove this bit of
irritating m a g i c that is very uncharacteristic from that usually
very engineering-driven dev team.

Being just about as old as I am, Emacs is even worse.  I'm changing my
config (in my separate .emacs.d repo) to to "10.5pt" "monospace."  For
whatever hysterical reason, Emacs font sizes are a hair smaller than
what everything else uses.