Create pavol script that handles volume/mute
The big 'advantage' is that I can no longer crank my volume to
dangerous levels anymore ^_~ because I have a check to see if the
volume is 120% or more (and arguably that is 'dangerous enough')

Also, pactl is really low-level and requires some custom handling, as
can be seen in the i3blocks-pulseaudio script.  Better to put that in
its own module.

Both i3 and i3blocks are rewired to use pavol rather than direct
manipulation of pactl.