Replaced defaults with nicer GTK/icons
The GTK theme is now Flat Remix (https://drasite.com/flat-remix-gtk).
Should the theme not pick up correctly, you'd want to use the Blue
Dark variant of Flat Remix.  The Flat Remix icons are way too
monochromatic for my tastes, so...

The icon theme is Numix Circle.  I briefly evaluated the Numix GTK
theme as well, but it's not a dark theme, and orange-red as the
highlight color is not to my taste.

Both fit in much better with the a e s t h e t i c than anything that
comes preinstalled on Ubuntu... but note that my artistic eye is kinda
weak, and I just went for things that were easily installable and
practical for daily use over obsessive pixel peeping