Install qute-keepassxc userscript
See here for this script's repository:
https://github.com/ususdei/qute-keepassxc

I'm copying it in directly because I've looked over the code and it
looks 'fine' for something that talks to my password store, and this
is the kind of code I'd like to keep within *my* control.

Note that there needs to be some config in keepassxc, and some
additional GPG configuration.  I hope no one else is using this code,
because obviously my email address will have to be changed!