Added i3blocks-todo script
At the moment it's super dumb... reads the topmost nonempty line and
prints it.

I may get delusions of grandeur and try to have it read my org-agenda
or something.  Not at the moment though; it's intended to be as dumb
and visible as possible to get me to do what's in it.