Install SVN prompt stuff in starship
This actually makes it a good bit slower now!  I actually had to sit
back and think about how best to integrate this so that it doesn't
unacceptably slow down everything.

The problem is that Starship is stupid fast by default, and so rando
hackjob shell scripts slow it down a lot.  svn info is no speed demon
either (about 50ms to figure out if I'm in a working copy or not).  So
I have a script that simply checks to see if the dir or its parents
have a .svn dir, easy way to check and it's "fast enough" for now.

The right way would be for Starship to get support for more VCSes than
just git.