Install/configure Subversion
I'm... pretty unenthusiastic about Subversion thus far.

Yes, it's a lot simpler than Git, and in my few minutes of using it,
I've generally not had to look up 'easy' things.  That is one of the
places Git falls on its face, yes.

The problem is

1. I've spent the better part of a decade memorizing Git's holes in
   usability, so "ease of use" isn't so much of a selling point

2. There's a whole host of places where svn makes things really
   painful for folks used to Git and the tooling around it, whereas
   Git needs almost no configuration

By default, Git paginates too-long diffs.  It is also trivial to
get it to colorize diffs.

SVN does none of this.

To get it to colorize diffs, you need to switch out its crap
"default" diff for your system diff, and tell it to use the color
flag.  Ok, that's not hard.

To get it to paginate, easier to just throw that away and make a shell
alias that calls vim.  That's not hard either.

But let's say we want to automatically symlink my changed
.subversion/config (y'know, to sync my dotfiles everywhere)... you
can't do this?  There's some "magic" somewhere that detects that I
deleted .subversion/config and puts back a default one.  I DON'T WANT
THAT!!!!!!111  (Read in this voice:
https://www.youtube.com/watch?v=_bW4vEo1F4E&t=107s)