Defenestrate treemacs
    defenestrate (n.):  To yeet (esp. out of a window).

Y'know, I'm really not getting *that* much out of treemacs that I
should force myself to play along with their workspace concept.  It's
not that it's bad --- if I'd started emacs with spacemacs/doom, I
probably would be used to it --- but more that it just doesn't work
with how I access projects, and so it all came into a high-speed
collision where I found myself feeling 300-400ms latency on keystrokes
inside PHP projects due to lsp-treemacs pulling in every darn project
on my filesystem because every time I projectile-switch-project I
added that project into a workspace even though I was curating those
workspaces.

And I'd really rather just not use it rather than have to curate my
workspace config like it was a bonsai tree, or build whatever elisp is
necessary to keep different workspaces from colliding.  Just not
enough time.

Argh!  I'm sorry to the treemacs people, it's a good project with a
lot of work put into it, but... it just wasn't meant to be.

dired is built in, and works fine.  It isn't a sidebar, but that's ok.