Fix org-deadline/org-refile not working
These definitions *used* to work, so either emacs got more strict
about function args or something changed with the advising/advised
functions.

Many thanks to everyone in the following question:
https://emacs.stackexchange.com/questions/52808/org-mode-cant-add-deadline-to-headline-wrong-number-or-arguments

This was the exact problem I was running into.  I suspect somewhere
someone suggested to advise various org manipulation commands with
org-save-all-org-buffers, and we all copied that into our .emacs and
thought very little of it.