Prevent pdf-tools-install on package install
The reasoning is that I always start Emacs in server mode, and so
it'll prompt for my sudo password... in some buffer I can't view,
because it's starting up in server mode...

After installation, running M-x pdf-tools-install gets it all set up.
I'd like to hook into this so that pdf-tools-install is run upon first
attempting to open a PDF, but I'm not there yet and it's not
*that* important to me.

Maybe look to see how Doom does it.