Remove lsp-ui and lsp hydra
The hydra wasn't really getting used in an effective way.

lsp-ui could be a speed hog, and I'm not really making effective use
of it either.

I've experienced performance problems with lsp-mode in general.  A lot
of it is this weird concept of a "workspace" that I cannot get my head
around.  I would be really happy if lsp-mode had the smarts to 1. not
add more projects into a "superworkspace" of all my projects 2. keep
all my projects separate so I don't have 28345279349837894373 files
open in LSP and 3. somehow tell the system to seamlessly switch
workspaces when I switch files.

---

I attempted to switch to eglot but it is broken when you install via
straight.el.  In general the eglot dev has taken the stance of "is it
in the Emacs repo or likely to go there?  if so, then it is good,
otherwise not."  This is a reasonable stance but also makes it not
work well with nonstandard things like straight.el (and, infamously,
flycheck).

For the time being I'll try to wrangle lsp-mode into MY conception of
how it should work.  If I fail... I'm going to hang onto my eglot code
in a separate branch and will probably maintain it as best I can.