Moved around many i3 keybindings for multihead
Reflecting experience with a computer with two monitors, I found I
needed more ability to move between monitors ('outputs') than I had
previously.

Lost the "switch workspace unconditionally" binding because my brain
doesn't work that way.  _on_output works much closer.

Currently mod+, and mod+. are disabled because I rarely used those
either and I'd rather have output switching keys instead.

mod+m used to be _m_aximize; now mod+f is _f_ullscreen (maximize), and
mod+shift+f is _f_loating (which I use a lot less).

---

It's been awhile since I committed to this repo.  I was in the process
of switching to home-manager and Nix, and there's a part of me that
still likes the idea of Nix.  But the Nix language is like alll the
worst parts of bash and Haskell in one yucky pile.  It says a lot about
what Nix promises that it remains compelling despite that.  Its
documentation, while still getting better, is also not great.

The end result is that I have a pile of Nix that sets up a lot of
command-line tools, but very little GUI tools (nvim-qt) because for
whatever reason home-manager and user-level Nix can't talk to my
Ubuntu's graphics stack, and so a *lot* of GUI utilities crash.

I would switch to Guix, but I'm not a fan of GNU.  I appreciate their...
fervor for what is right (presuming it's not regarding their Dear
Leader), but sometimes I wanna play a game or talk to some proprietary
thing, and the moment that happens, GNUism stops looking like a
principled stand against creeping corporatism and more like impractical
software veganism.

I also have the problem that I tend to overcustomize my environments.
I'm actively working on reducing my need for customizations.

I may yet deprecate this repo and replace it with a bunch of Nix, but
the jury is out yet...