i3: Substantial changes
I moved hsplit/vsplit from -| to sv like in vim.

I also moved a bunch of keys around for the sake of moving them around.

And the below is my goal.....

I'm attempting to move away from all-singing, all-dancing terminal
emulators like kitty, and towards something simpler, probably alacritty.
I'm going to say 'kitty' a bunch as a representative of this genre of
programs.    This is not to pick on kitty, I've used it a lot over the
past 3-4 months, I like it, and it has nice features.  Similar problems
also exist in related programs like WezTerm and Terminator, as well as
tmux itself.

Keep in mind that I'm not focusing on my terminal setup every minute of
every day, so I may be somewhat off base on the below.  If so, I
apologize.

I have (or at least perceive) these problems:

 - I don't like needing three bucky bits:  control for my program, super
   for my WM, and meta for kitty and things like tmux (which, yes, I do
   not use tmux in kitty, that's usually pointless).  It's confusing to
   remember whether this pane is in kitty or is an i3 window, for me. i3
   has the capability to arrange windows in panes and in tabs, and to
   search them with something like rofi, why not use that?

 - Once you remove that, all kitty/WezTerm offer over something
   radically simpler like alacritty is... things like images.  I like
   images, but that is still very much in flux, and I'd like things to
   settle down a bit before committing to a terminal with image support.

 - These are newish terminals, and it shows.  kitty was very stable,
   but it always took about 2-3 seconds to load on my work machine.
   This is fine if I was going to run one instance of kitty and
   multiplex; but when I'm running many of them?  WezTerm must be using
   some OpenGL/etc stuff that pisses off my computer; my Firefox always
   feels slower when I'm running it, but not when using kitty or
   alacritty.  Go figure.  Then again, I can barely run Google Maps in
   same Firefox; I'm hoping a Linux update fixes this.

If I wasn't running a tiling WM already, the value of running something
like kitty or wezterm goes WAAAY up, and I hope the authors of the
various terminals keep working at it, I'd like to show graphics in
terminals eventually.